requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
transformers
PyPDF2
torch 
jinja2